backend/pyproject.toml
[project]
name = "doctor-cornelius"
version = "0.1.0"
description = "Backend for Doctor Cornelius, the Trusted Archivist of Narnia Labs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.125.0",
    "uvicorn>=0.38.0",
    "slack-sdk>=3.33.0",
    "slack-bolt>=1.21.0",
    "graphiti-core>=0.9.0",
    "google-genai>=1.0.0",
    "neo4j>=5.25.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "apscheduler>=3.10.0",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "aiohttp>=3.13.2",
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "pyrefly>=0.16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/doctor_cornelius"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pyrefly]
project_includes = ["src"]
python_version = "3.11"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
